+++ src/has-magic.ts
@@ -7,14 +7,8 @@ export const hasMagic = (
   if (!Array.isArray(pattern)) {
     pattern = [pattern]
   }
-  return pattern.some(p => {
-    const g = new Glob(p, options)
-    if (g.matchSet.length === 0) {
-      return false
-    }
-    if (g.matchSet.length > 1) {
-      return true
-    }
-    return g.matchSet[0].some(p => typeof p !== 'string')
-  })
+  const g = new Glob(pattern, options)
+  return g.patterns.length === 0
+    ? false
+    : g.patterns.some(p => p.hasMagic())
 }